Make it easier to add a book
This swaps out the existing ISBN-driven creation flow in favour of a 
more flexible form, allowing users to fill out all the book attributes 
as appropriate. The ISBN-lookup feature now exists as an optional call
which can be made from the form.

This also means that the ISBN lookup behaviour is no longer coupled to 
the Book model.