Add support for shelves
Two shelves will be created by the default migration:

- 'Third floor'
- 'Sixth floor'

A shelf can be set for a book when it's not out on loan.

When a book is returned from loan, the shelf is also specified.

A record of the shelf which the book is returned to is kept in the
`loans` table.