Fix failing tests
Two tests are failing with Capybara::Ambiguous match errors - Capybara
is finding more shelves in the shelf drop down menu when on the page to
return a book than expected. This is because we use FactoryGirl to create
shelves for the tests, but the test database already contains shelves - we
already have shelves in the test database from the shelf data migration.

Since the test database already contains shelves, this commit stops
creating more shelves with factories.