Pin omniauth-google-oauth2 gem to use later version
The omniauth-google-oauth2 gem had no version in the Gemfile, so we were
using version 0.4.1. This version relies on the Google+ API, which  will
stop working on 7/3/19, but will start to intermittently fail from 28/1/19.
We need to be using version 0.5.4 or higher of omniauth-google-oauth2 to hit
the new Google OAuth endpoints.