Upgrade Rails to 5.2.2
Upgraded Rails to 5.2.2 and ran `rails app:update` to get the changes
made in Rails 5.2.

- Many of the changes were fairly minor.
- I added some new files and settings even though we're not currently
using them and the code is commented out to make things clearer is we
need to use these in future .
- If we want to use all the changes in `new_framework_defaults.rb` and
`new_framework_defaults_5_2.rb` we could remove these files and change
`config/application.rb` to contain `config.load_defaults 5.2`. However a
couple of these changes are not compatible with backwards versions of
Rails, so it's best to check everything's working before changing these.