Fix the issue with wrongly displaying book covers
We were checking `book.google_id`and if that was true using the
book URL from Google. The `google_id` column now seems to contain a mix
of `null`s and empty strings. When `google_id` is an empty string,
`book.google_id` is true, so we try and get a book image from Google
when we shouldn't.

This change uses the `.present?` method to check for both `nil` and
empty strings.

Tests to follow.