Fix finding books by ISBN
Passing `isbn` as a keyword argument to both
`BookMetadataLookup.find_by_isbn` and `Openlibrary::Data.find_by_isbn`
no longer works, so this changes the code to pass isbn as a positional
argument in both cases.

We can also remove the empty options hash and the `ENV["REQUEST_IP"]`
arguments from `GoogleBooks.search`. `REQUEST_IP` is no longer set as an
environment variable - this was only needed when deploying to Heroku.