Upgrade Omniauth to fix a security vulnerability
This upgrades the `omniauth-google-oauth2` gem to bring in version 2 of
omniauth. In order to work with the new version, the request to
'/auth/google' now has to be a POST request. The
`omniauth-rails_csrf_protection` gem is needed "Because Rails handles its
CSRF protection in its RequestForgeryProtection class, and stores tokens
in a non-vanilla-rack friendly way, you must pass a rails-friendly validator
in instead" (https://github.com/omniauth/omniauth/wiki/Upgrading-to-2.0).