Fix bug with getting the author from openlibrary
When adding a new book and choosing to look up its details from the ISBN
number, we were not setting the book's author correctly when getting the
author from openlibrary.

The `authors` method on the openlibrary data that is returned is not the
authors name, but it is an array of hashes containing author details.
This fixes the code so that it gets the author's name from the data, and
doesn't set the author name to be the array.

If there are multiple authors, the names get formatted to match the way
that Google formats multiple authors: "Author One, Author Two, Author
Three".