Add test for UserController
It's a very basic test, but it adds a bit of test coverage where there
was none before.

I've kept the existing pattern of the controller tests even though the
Rails way is now to write controller tests as functional tests. This is
for consistency, but all the controller tests can be revisted later.