Change column type of User image_url from String to Text
Because it is possible for people to have a Google image URL that is
longer than the 255 character limit that was in place. If they did have
a long image_url, they would see an uncaught error when trying to log
in, since the URL cannot be saved or updated in the database.