Exclude migration file from Rubocop check
Rubocop was warning about the migration to change the Users image_url
data type from string to text, which was not written in a reversible
way. This adds a line to the Rubocop config to ignore the warning since
the migration should not need to be reversed, and writing it in a
reversible way would not be particuarly straightforward because we may
have data in that column which is too long to be changed back to a
string type.