Upgrade Rails from 6.1.4.7 to 7.0.2.3
This is the result of upgrading the Rails gem and running the `rails
app:update` task. A few minor changes were needed to get things
working, and not all of the new settings in `new_framework_defaults_7_0.rb`
have been applied yet - the remaining ones will be checked or changed
later.