Use all defaults from Rails 7
There were a couple of new Rails 7 features that hadn't been turned on
when the app was upgraded to Rails 7. This turns on the remaining
Rails 7 settings, meaning we can delete the
`new_framework_defaults_7_0.rb` file since all the settings are applied.

The only change which does have a noticeable affect is the change of the
`key_generator_hash_digest_class` from SHA1 to SHA256. This change means
that everyone will need to log in again when next using the app since
the session cookies will be encrypted differently. It would be possible
to build a key rotator to avoid this (and the upgrade docs do suggest
this). However since the app hardly used at the moment, the key rotator
would be need to be in place for a very long time and making people log
in again shouldn't cause them any issues.

Changing `self.abstract_class = true` to `primary_abstract_class` in
`app/models/application_record.rb` was just to change to the method that
new apps use, but makes no difference to how things work.