Standardise test styles slightly by removing should syntax
The `minitest-spec-rails` gem has the option to use allow the
[Shoulda](https://github.com/thoughtbot/shoulda) methods in tests
(`context`, `should` and `should_eventually`). According to the docs,
this is designed to help when migrating away from Shoulda.

The tests are in a mix of styles, so this change aims to simplify them a
little by just using `describe` and `it` everywhere consistently instead
of having a mix of `describe` / `context` and `should` / `it`.