Write RootController tests in a functional test format
This rewrites the RootController tests to use a functional test format.
Doing this showed the issue with the existing tests - we were testing
that an instance variable gets set in the controller, but we actually
stopped using the instance variable in the view 8 years ago (in
5f2cbaddf6014074e509870b8e84348db571b5cb). I haven't
replaced all the tests here - some will need to be added elsewhere.

As well as updating the tests, this also deletes the unused instance
variable from the controller and added an path helper for the library
csv route.