Add more tests for the Books controller
Running a test coverage tool against the repo showed that the level of test
coverage for the books controller was poor, so this adds more tests to
try and fix the issue. The main areas missing tests were related to creating
and updating books - this adds tests for when these things happen
successfully as well as for the error messages when a book can't be
created or updated.