Remove shelves from the loans
We don't have two shelves any more, so will need to change how we deal
with shelves / locations of copies. As a first step, this removes the
concept of shelves from loans. A loan now no longer belongs to a shelf,
which means that we don't have to keep a record of shelves which no
longer exist - we only care about current shelves and where a copy is now.