Rename "shelves" to "locations"
We want the app to show copies in different locations, but the different
locations are going to be different office buildings, and not different shelves
in the same building. To prepare for this change, this commit starts to
rename "shelf" and "shelves" to "location" and "locations" in the
backend code. This renames

- The `shelves` table to `locations`
- The `shelf_id` column of the `copies` table to `location_id`
- Method names, and HTML class and ID names