Fix bug with `formatted_version_author` helper method
We try to look up a user by ID if the ID exists. This should be fine
but it is sometimes giving errors because of the state of the data -
we have some user IDs in the `versions` table for users who no longer
exist in the `users` table.

This fixes the issue by no longer making the assumption that
`User.where(id: user_id).first` will return a User when passed a
user_id.