Remove unecessary `<nav>`s
The book and copy pages were using a `<nav>` element for links for
certain actions. These links are not really navigation links, so this
changes them to a `<div>` instead of adding a different label for each
`<nav>` on the page (which we'd need to do if we were to keep both).