Remove `<h1>` elements from nav
The navigation had two `<h1>` elements - the library title, and the user's
name and loan count. Neither of these should be an `<h1>`, and we were
ending up with 3 `<h1>` elements on some pages as a result.

This does not attempt to change any styling, but instead replaces just
uses `<span>`s to add the existing styles.