Fix headings on book index page
This page had an `<h3>` which should not have been an `<h3>`, and no
`<h1>`. I've removed the h3 element and replaced it with a div. The
style looked different, despite updating the relevant place in the
stylesheet, because we were getting some of the styles from the user
agent stylesheet since there's no reset. I've added a few styles to keep
things looking similar, but would like to revisit this later.

This also adds an h1 since the page didn't have one.