Make `book_reference` column of copy nullable
This is stage one of the work to get rid of the `book_reference` column
from the copies table. This change removes the constraints from the
column so that we can stop writing to it later. There's no change to the
behaviour now, since we are still writing to the column.