Use `copy_id` instead of `book_reference` everywhere
Previous commits stopped saving `book_reference` and made it a
non-required field. Now we stop using it altogether, and switch to
always using `copy_id`. So the paramter when viewing a copy is now the
`copy_id` and not the `book_reference` and every time we look up a copy
it is by it's id.