Add tests for copies controller
There was no test file specifically for the copies controller but bits
of its functionality were tested in other tests, mainly the copy actions
integration tests. This adds a test file for the copies controller to
ensure we have full test coverage for it. There is a bit of duplication
with some of the of tests in copy_actions_test.rb, but I have left this
since the integration tests are using Capybara (unlike these new tests)
and the plan is to revisit the integration tests.