Stop linking to book history pages
We decided that being able to view the book history pages isn't
particularly useful for people - no one is likely to be interested in
whether the title of author of a book was changed slightly. This removes
the link, but keeps the pages and the data since the pages may still be
useful for people managing the library.