Remove all references to location
We currently only have one location for the library, so want to simplify
the site and avoid confusion by removing all references to locations
(for example filtering by locations, setting a location). This keeps the
model associations but removes everything else. If we do decide to add
locations it is not guaranteed that they will work the same way since
they would likely be in different cities and not just different floors
of a building.